requirements.txt
torch
torchvision
torchaudio

mlflow
dvc
evidently
streamlit

pandas
numpy
scikit-learn
matplotlib
pyyaml
Pillow